requirements.txt
tensorflow==2.12.0
transformers==4.30.2
joblib
numpy
pandas
scikit-learn
google-generativeai
gradio